API

MultiResponseVarianceComponentModels.MultiResponseVarianceComponentModels

MultiResponseVarianceComponentModels.MRVCModel

MultiResponseVarianceComponentModels.fisher_B!

MultiResponseVarianceComponentModels.fisher_Σ!

MultiResponseVarianceComponentModels.fit!

MultiResponseVarianceComponentModels.h2

MultiResponseVarianceComponentModels.kron_axpy!

MultiResponseVarianceComponentModels.kron_reduction!

MultiResponseVarianceComponentModels.loglikelihood!

MultiResponseVarianceComponentModels.loglikelihood_miss!

MultiResponseVarianceComponentModels.lrt

MultiResponseVarianceComponentModels.permute

MultiResponseVarianceComponentModels.rg

MultiResponseVarianceComponentModels.update_B!

MultiResponseVarianceComponentModels.update_B_miss!

MultiResponseVarianceComponentModels.update_Γk!

MultiResponseVarianceComponentModels.update_Σ!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.update_Σk_miss!

MultiResponseVarianceComponentModels.MultiResponseVarianceComponentModels — Module

MultiResponseVarianceComponentModels.jl or MRVCModels.jl permits maximum likelihood (ML) and residual maximum likelihood (REML) estimation as well as inference for multivariate response variance components linear mixed models.

MultiResponseVarianceComponentModels.MRVCModel — Method

MRVCModel(Y, X, V)

Create a new MRVCModel instance from response matrix Y, predictor matrix X, and kernel matrices V.

Positional arguments

Y::AbstractVecOrMat : response matrix

X::Union{Nothing, AbstractVecOrMat} : predictor matrix

V::Union{AbstractMatrix, Vector{<:AbstractMatrix}} : kernel matrices

Keyword arguments

se::Bool : calculate standard errors. Default is true.

reml::Bool : REML estimation instead of ML estimation. Default is false.

MultiResponseVarianceComponentModels.fisher_B! — Method

fisher_B!(model::MRVCModel)

Compute the sampling variance-covariance of regression coefficients model.B, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.fisher_Σ! — Method

fisher_Σ!(model::MRVCModel)

Compute the sampling variance-covariance of variance component estimates model.Σ, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.fit! — Method

fit!(model::MRVCModel)

Fit a multivariate response variance components model by MM or EM algorithm.

Keyword arguments

maxiter::Integer : maximum number of iterations. Default is 1000.

reltol::Real : relative tolerance for convergence. Default is 1e-6.

verbose::Bool : display algorithmic information. Default is true.

init::Symbol : initialization strategy. :default initialize by least squares. :user uses user supplied value at model.B and model.Σ.

algo::Symbol : optimization algorithm. :MM (default) or EM.

log::Bool : record iterate history or not. Defaut is false.

Output

history : iterate history.

MultiResponseVarianceComponentModels.h2 — Method

h2(model::MRVCModel)

Calculate heritability estimates and their standard errors, assuming that all variance components capture genetic effects except the last term. Also return total heritability from sum of individual contributions and its standard error.

MultiResponseVarianceComponentModels.kron_axpy! — Method

kron_axpy!(A, X, Y)

Overwrite Y with A ⊗ X + Y. Same as Y += kron(A, X), but more memory efficient.

MultiResponseVarianceComponentModels.kron_reduction! — Method

kron_reduction!(A, B, C; sym = false)

Overwrite C with the derivative of tr(A' (X ⊗ B)) wrt X. C[i, j] = dot(Aij, B), where Aij is the (i , j) block of A. sym = true indicates A and B are symmetric.

MultiResponseVarianceComponentModels.loglikelihood! — Method

loglikelihood!(model::MRVCModel)

Overwrite model.storage_nd_nd by inverse of the covariance matrix model.Ω, overwrite model.storage_nd by U' \ vec(model.R), and return the log-likelihood. Assume model.Ω and model.R are already updated according to model.Σ and model.B.

MultiResponseVarianceComponentModels.loglikelihood_miss! — Method

loglikelihood_miss!(model::MRVCModel)

Overwrite model.storage_nd_nd by inverse of the covariance matrix model.Ω, overwrite model.storage_nd by U' \ vec(model.R), overwrite model.storage_n_miss_n_miss_1 by conditional variance, and return the surrogate Q-function of log-likelihood. Assume model.Ω and model.R are already updated according to model.Σ and model.B.

MultiResponseVarianceComponentModels.lrt — Method

lrt(model1::MRVCModel, model0::MRVCModel)

Perform a variation of the likelihood ratio test for univariate variance components models as in Molenberghs and Verbeke 2007 with model1 and model0 being the full and nested models, respectively.

MultiResponseVarianceComponentModels.permute — Method

permute(Y::AbstractVecOrMat)

Return permutation P such that vec(Y)[P] rearranges vec(Y) with missing values spliced after non-missing values. Also return inverse permutation invP such that vec(Y)[P][invP] = vec(Y).

MultiResponseVarianceComponentModels.rg — Method

rg(model::MRVCModel)

Calculate genetic/residual correlation estimates and their standard errors.

MultiResponseVarianceComponentModels.update_B! — Method

update_B!(model::MRVCModel)

Update the regression coefficients model.B, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.update_B_miss! — Method

update_B_miss!(model::MRVCModel)

Update the regression coefficients model.B, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd and conditional covariance model.storage_n_miss_n_obs_1 precomputed.

MultiResponseVarianceComponentModels.update_Γk! — Method

update_Γk!(model, k)

Update the parameters model.Γ[k] and model.Ψ[:,k] assuming a diagonal plus low rank structure for model.Σ[k]. Assumes covariance matrix model.Ω is available at model.storage_nd_nd and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σ! — Method

update_Σ!(model::MRVCModel)

Update the variance component parameters model.Σ, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd. For missing response, assume conditional variance model.storage_n_miss_n_miss_1 is precomputed.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model, k, rk)

Update the model.Σ[k] assuming it has rank rk < d, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model::MRVCModel, k, Val(:EM))

EM update the model.Σ[k] assuming it has full rank d, inverse of covariance matrix model.Ω is available at model.storage_nd_nd, and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model::MRVCModel, k, Val(:MM))

MM update the model.Σ[k] assuming it has full rank d, inverse of covariance matrix model.Ω is available at model.storage_nd_nd, and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σk_miss! — Method

update_Σk_miss!(model::MRVCModel, k, Val(:MM))

MM update the model.Σ[k] assuming it has full rank d, inverse of covariance matrix model.Ω is available at model.storage_nd_nd, and model.Ω⁻¹R, conditional variance model.storage_n_miss_n_miss_1, and Ω⁻¹P'CPΩ⁻¹ precomputed.