API

MultiResponseVarianceComponentModels.fisher_B!

MultiResponseVarianceComponentModels.fisher_Σ!

MultiResponseVarianceComponentModels.fit!

MultiResponseVarianceComponentModels.h2

MultiResponseVarianceComponentModels.kron_axpy!

MultiResponseVarianceComponentModels.kron_reduction!

MultiResponseVarianceComponentModels.loglikelihood!

MultiResponseVarianceComponentModels.lrt

MultiResponseVarianceComponentModels.rg

MultiResponseVarianceComponentModels.update_B!

MultiResponseVarianceComponentModels.update_Γk!

MultiResponseVarianceComponentModels.update_Σ!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.update_Σk!

MultiResponseVarianceComponentModels.fisher_B! — Method

fisher_B!(model::MultiResponseVarianceComponentModel)

Compute the sampling variance-covariance of regression coefficients model.B, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.fisher_Σ! — Method

fisher_Σ!(model::MultiResponseVarianceComponentModel)

Compute the sampling variance-covariance of variance component estimates model.Σ, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.fit! — Method

fit!(model::MultiResponseVarianceComponentModel)

Fit a multivariate response variance component model by an MM or EM algorithm.

Positional arguments

model : a MultiResponseVarianceComponentModel instance.

Keyword arguments

maxiter::Integer : maximum number of iterations. Default is 1000.

reltol::Real : relative tolerance for convergence. Default is 1e-6.

verbose::Bool : display algorithmic information. Default is false.

init::Symbol : initialization strategy. :default initialize by least squares. :user uses user supplied value at model.B and model.Σ.

algo::Symbol : optimization algorithm. :MM (default) or EM.

log::Bool : record iterate history or not. Defaut is false.

reml::Bool : REML instead of ML estimation. Default is false.

se::Bool : calculate standard errors. Default is true.

Output

history : iterate history.

MultiResponseVarianceComponentModels.h2 — Method

h2(model::MultiResponseVarianceComponentModel)

Calculate heritability estimates and their standard errors, assuming that all variance components capture genetic effects except the last term. Also returns total heritability and its standard error from sum of individual contributions.

MultiResponseVarianceComponentModels.kron_axpy! — Method

kron_axpy!(A, X, Y)

Overwrite Y with A ⊗ X + Y. Same as Y += kron(A, X), but more memory efficient.

MultiResponseVarianceComponentModels.kron_reduction! — Method

kron_reduction!(A, B, C; sym = false)

Overwrite C with the derivative of tr(A' (X ⊗ B)) wrt X. C[i, j] = dot(Aij, B), where Aij is the (i , j) block of A. sym = true indicates A and B are symmetric.

MultiResponseVarianceComponentModels.loglikelihood! — Method

loglikelihood!(model::MultiResponseVarianceComponentModel)

Overwrite model.storage_nd_nd by inverse of the covariance matrix model.Ω, overwrite model.storage_nd by U' \ vec(model.R), and return the log-likelihood. This function assumes model.Ω and model.R are already updated according to model.Σ and model.B.

MultiResponseVarianceComponentModels.lrt — Method

lrt(model1::MultiResponseVarianceComponentModel, model0::MultiResponseVarianceComponentModel)

Perform a variation of the likelihood ratio test for univariate variance components models as in Molenberghs and Verbeke 2007 with model1 and model0 being the full and nested models, respectively.

MultiResponseVarianceComponentModels.rg — Method

rg(model::MultiResponseVarianceComponentModel)

Calculate genetic/residual correlation estimates and their standard errors.

MultiResponseVarianceComponentModels.update_B! — Method

update_B!(model::MultiResponseVarianceComponentModel)

Update the regression coefficients model.B, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.update_Γk! — Method

update_Γk!(model, k)

Update the parameters model.Γ[k] and model.Ψ[:,k] assuming a diagonal plus low rank structure for model.Σ[k]. Assumes covariance matrix model.Ω is available at model.storage_nd_nd and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σ! — Method

update_Σ!(model::MultiResponseVarianceComponentModel)

Update the variance component parameters model.Σ, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model, k, rk)

Update the model.Σ[k] assuming it has rank rk < d, assuming inverse of covariance matrix model.Ω is available at model.storage_nd_nd and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model::MultiResponseVarianceComponentModel, k, Val(:EM))

EM update the model.Σ[k] assuming it has full rank d, inverse of covariance matrix model.Ω is available at model.storage_nd_nd, and model.Ω⁻¹R precomputed.

MultiResponseVarianceComponentModels.update_Σk! — Method

update_Σk!(model::MultiResponseVarianceComponentModel, k, Val(:MM))

MM update the model.Σ[k] assuming it has full rank d, inverse of covariance matrix model.Ω is available at model.storage_nd_nd, and model.Ω⁻¹R precomputed.